refactor: Runtime: Move create and state into default impl
