+++ runtime/src/runtime/fvm.rs
@@ -36,8 +36,6 @@ use crate::runtime::{
 };
 use crate::{actor_error, ActorError, Runtime};
 
-use super::EMPTY_ARR_CID;
-
 /// A runtime that bridges to the FVM environment through the FVM SDK.
 pub struct FvmRuntime<B = ActorBlockstore> {
     blockstore: B,
@@ -233,27 +231,6 @@ where
                     }
                 }
             })
-    }
-
-    fn create<T: Serialize>(&mut self, obj: &T) -> Result<(), ActorError> {
-        let root = fvm::sself::root()?;
-        if root != EMPTY_ARR_CID {
-            return Err(
-                actor_error!(illegal_state; "failed to create state; expected empty array CID, got: {}", root),
-            );
-        }
-        let new_root = ActorBlockstore.put_cbor(obj, Code::Blake2b256)
-            .map_err(|e| actor_error!(illegal_argument; "failed to write actor state during creation: {}", e.to_string()))?;
-        fvm::sself::set_root(&new_root)?;
-        Ok(())
-    }
-
-    fn state<T: DeserializeOwned>(&self) -> Result<T, ActorError> {
-        let root = fvm::sself::root()?;
-        Ok(ActorBlockstore
-            .get_cbor(&root)
-            .map_err(|_| actor_error!(illegal_argument; "failed to get actor for Readonly state"))?
-            .expect("State does not exist for actor state root"))
     }
 
     fn get_state_root(&self) -> Result<Cid, ActorError> {
